Update README.md
This movie is so nice and that's good for me. they have also a good moral like Cameron said, 'Avatar asks us to see that everything is connected, all human beings to each other, and us to the Earth. ' This is a clear statement of religious belief. This is pantheism.